backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
click==8.3.2
fastapi==0.136.1
greenlet==3.4.0
h11==0.16.0
hdbscan==0.8.42
httptools==0.7.1
idna==3.13
iniconfig==2.3.0
joblib==1.5.3
mypy_extensions==1.1.0
numpy==2.4.4
packaging==26.0
paho-mqtt==2.1.0
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
scikit-learn==1.8.0
scipy==1.17.1
SQLAlchemy==2.0.49
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
psycopg[binary]